% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentAnalysis.R
\name{contentAnalysis}
\alias{contentAnalysis}
\title{Perform Content Analysis}
\usage{
contentAnalysis(articles_df, k, words_to_remove)
}
\arguments{
\item{articles_df}{Dataframe of articles. Output from getArticles() function.}

\item{k}{Number of topics}

\item{words_to_remove}{(Optional) Additional stopwords to be removed}
}
\value{
List
}
\description{
Perform content analysis; build LDA model on abstract text
}
\examples{
content <- contentAnalysis(articles_df =  articles_df, k = 10)

# Add your own stop words
words_to_remove <- c("stopword1", "stopword2")
content <- contentAnalysis(articles_df =  articles_df, k = 10, words_to_remove = words_to_remove)

# Word cloud
wordCloud(content)

# Top words in corpus
topNWordsInDoc(content, n = 10)

# Plot top words
topNWordsInDoc(content, n = 10, plot = TRUE)

# Top n words in topics
topNWordsInTopic(content, n = 3)

# Plot top n words in topics
topNWordsInTopic(content, n = 3, plot = TRUE)

}
