% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArticles.R
\name{getArticles}
\alias{getArticles}
\title{Load and process articles CSV files.}
\usage{
getArticles(files_path, data_source = "Scopus")
}
\arguments{
\item{files_path}{Path to files.}

\item{data_source}{Data source..}
}
\value{
dataframe.
}
\description{
Load and process articles CSV files.
}
\examples{
articles_df <- getArticles(files_path = read_data, data_source = data_source)

}
