% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactAnalysis.R
\name{impactAnalysis}
\alias{impactAnalysis}
\title{Load and process articles CSV files.}
\usage{
impactAnalysis(articles_df)
}
\arguments{
\item{files_path}{Path to files.}

\item{data_source}{Data source..}
}
\value{
dataframe.
}
\description{
Load and process articles CSV files.
}
\examples{
articles_df <- getArticles(files_path = read_data, data_source = data_source)

}
