% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureAnalysis.R
\name{structureAnalysis}
\alias{structureAnalysis}
\title{Perform Structure Analysis}
\usage{
structureAnalysis(impact_object)
}
\arguments{
\item{articles_df}{Path to files}
}
\value{
List
}
\description{
Perform structure analysis on impactCLR object
}
\examples{
structure <- structureAnalysis(impact_object = impact)

# Components vs Size
plot(structure, loess = T)

# Plot structure node : edges
showGraphs(structure)

# Plot components
graphComponents(structure, n = 5)

}
