% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStrFiles.R
\name{writeStrFiles}
\alias{writeStrFiles}
\title{Perform Structure Analysis}
\usage{
writeStrFiles(structure_list, impact_list, gml_name, txt_name, path = getwd())
}
\arguments{
\item{structure_list}{structureCLR object}

\item{impact_list}{impactCLR object}

\item{gml_name}{gml file name}

\item{txt_name}{Output UCINET file name}

\item{path}{Output path}
}
\value{
NA
}
\description{
Perform Structure Analysis
}
