% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Fit a CLR via dimension reduction}
\usage{
clr(Y, X, clust, qx_estimation = "pctvar", ortho_Y = TRUE,
  qy_estimation = "pctvar", d_estimation = "cor", pct = 0.999)
}
\arguments{
\item{pct}{}
}
\value{

}
\description{
blabla
}
\examples{
library(clr)

Sys.setenv(TZ = 'Europe/London')
PB between High Sierra and R
data(gb_load)

clr_load <- clrdata(x = gb_load$ENGLAND_WALES_DEMAND,
                    order_by = gb_load$TIMESTAMP,
                    support_grid = 1:48)

# data cleaning: replace zeros with NA
clr_load[rowSums((clr_load == 0) * 1) > 0, ] <- NA
matplot(t(clr_load), ylab = 'Daily loads', type = 'l')

Y <- clr_load[2:nrow(clr_load), ]
X <- clr_load[1:(nrow(clr_load)-1), ]
}
