% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Fit a CLR via dimension reduction}
\usage{
clr(y_data, x_data, clust, qx_estimation = "pctvar", scale_x = TRUE,
  ortho_y = TRUE, kj = NULL, qy_estimation = "pctvar",
  d_estimation = "cor", pct = 0.999)
}
\arguments{
\item{pct}{}
}
\value{

}
\description{
blabla
}
\examples{

}
