% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Fit a CLR via dimension reduction}
\usage{
clr(y, x, clust, qx_estimation = "pctvar", scale_x = TRUE, ortho_y = TRUE,
  kj = NULL, qy_estimation = "pctvar", d_estimation = "cor",
  pct = 0.999)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
blabla
}
\examples{
library(clr)

data(gb_load)

clr_load <- clrdata(x = gb_load$ENGLAND_WALES_DEMAND,
                    order_by = gb_load$TIMESTAMP,
                    support_grid = 1:48)

clr_weather <- clrdata(x = gb_load$TEMPERATURE,
                       order_by = gb_load$TIMESTAMP,
                       support_grid = 1:48)
}
