% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_data.R
\name{clrdata}
\alias{clrdata}
\title{Title}
\usage{
clrdata(x = NULL, order_by = index(x), support_grid)
}
\arguments{
\item{support_grid}{}
}
\description{
\code{clrdata} is used to create a \code{clrdata} object from raw data
inputs.
}
\examples{

}
