% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_test-data.R
\docType{data}
\name{clust_test}
\alias{clust_test}
\title{Electricity load example: clusters on test set}
\format{A list of length 14: \describe{
  14 clusters of loads, depending on both daily and seasonal classification,
  banking holidays being removed}}
\usage{
clust_test
}
\description{
A list with observations by cluster for prediction
}
\author{
Amandine Pierrot <amandine.m.pierrot@gmail.com>
}
\keyword{datasets}
