% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_train-data.R
\docType{data}
\name{clust_train}
\alias{clust_train}
\title{Electricity load example: clusters on train set}
\format{A list of length 14: \describe{
  14 clusters of loads, depending on both daily and seasonal classification,
  banking holidays being removed}}
\usage{
clust_train
}
\description{
A list with observations by cluster for fitting
}
\author{
Amandine Pierrot <amandine.m.pierrot@gmail.com>
}
\keyword{datasets}
