% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactAnalysis.R
\name{impactAnalysis}
\alias{impactAnalysis}
\title{Perform Impact Analysis}
\usage{
impactAnalysis(articles_df)
}
\arguments{
\item{articles_df}{Dataframe of articles. Output from getArticles() function.}
}
\value{
List
}
\description{
Perform impact analysis on articles dataframe
}
\examples{
impact <- impactAnalysis(articles_df = articles_df)
plot(impact)

# Top cited papers
topCites(impact, n = 10)

# Most common sources
topSources(impact, n = 6)

}
