% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-clr.R
\name{predict.clr}
\alias{predict.clr}
\title{Model predictions with Curve Linear Regression}
\usage{
\method{predict}{clr}(object, newX = NULL, clust = NULL, newXmean = NULL,
  ...)
}
\arguments{
\item{newXmean}{}
}
\description{
Model predictions with Curve Linear Regression
}
\examples{
library(clr)

Sys.setenv(TZ = 'Europe/London')
# PB between High Sierra and R
data(gb_load)

clr_load <- clrdata(x = gb_load$ENGLAND_WALES_DEMAND,
                    order_by = gb_load$TIMESTAMP,
                    support_grid = 1:48)

# data cleaning: replace zeros with NA
clr_load[rowSums((clr_load == 0) * 1) > 0, ] <- NA

Y <- clr_load[2:nrow(clr_load), ]
X <- clr_load[1:(nrow(clr_load) - 1), ]

begin_pred <- which(substr(rownames(Y), 1, 4) == '2016')[1]
Y_train <- Y[1:(begin_pred - 1), ]
X_train <- X[1:(begin_pred - 1), ]

model <- clr(Y = Y_train, X = X_train)

pred_on_train <- predict.clr(model)
head(pred_on_train[[1]])

X_test <- X[begin_pred:nrow(X), ]
pred_on_test <- predict.clr(model, newX = X_test)
head(pred_on_test[[1]])
}
